mock-sso/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt==2.8.0
jinja2==3.1.3
python-multipart==0.0.6
